[REFACTOR] LotteClient의 책임 분리
- LotteClient는 외부api와 http통신 및 body값 deserialize
- LotteClientService는 응답코드에 따른 예외처리
