+++ src/test/java/com/tkl/refactoring/chapter4/LotteClientTest.java
@@ -1,7 +1,6 @@
 package com.tkl.refactoring.chapter4;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.IOException;
 import java.util.List;
@@ -28,67 +27,18 @@ class LotteClientTest {
 									  .setBody(
 											  """
 											  [
-												  {
-													"id":"1",
-													"name":"Performance 1",
-													"startDate":"2023-01-01",
-													"endDate":"2023-01-10"
-												  },
-												  {
-													"id":"2",
-													"name":"Performance 2",
-													"startDate":"2023-02-01",
-													"endDate":"2023-02-10"
-												  }
-											  ]
-											  """
-											  )
-									  );
-
-		String baseUrl = "http://localhost:" + mockWebServer.getPort();
-		LotteClient lotteClient = new LotteClient(baseUrl);
-
-		// when
-		List<LotteClient.Performance> performances = lotteClient.getPerformances();
-
-		// then
-		assertThat(performances).hasSize(2)
-				.anySatisfy(performance -> {
-					assertThat(performance.getId()).isEqualTo("1");
-					assertThat(performance.getName()).isEqualTo("Performance 1");
-					assertThat(performance.getStartDate()).isEqualTo("2023-01-01");
-					assertThat(performance.getEndDate()).isEqualTo("2023-01-10");
-				})
-				.anySatisfy(performance -> {
-					assertThat(performance.getId()).isEqualTo("2");
-					assertThat(performance.getName()).isEqualTo("Performance 2");
-					assertThat(performance.getStartDate()).isEqualTo("2023-02-01");
-					assertThat(performance.getEndDate()).isEqualTo("2023-02-10");
-				});
-
-	}
-
-	@Test
-	void 응답코드가_500이면_예외가_떨어진다() {
-		// given
-		mockWebServer.enqueue(new MockResponse()
-									  .setResponseCode(HttpStatus.INTERNAL_SERVER_ERROR.value())
-									  .setHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-									  .setBody(
-											  """
-											  [
-												  {
-													"id":"1",
-													"name":"Performance 1",
-													"startDate":"2023-01-01",
-													"endDate":"2023-01-10"
-												  },
-												  {
-													"id":"2",
-													"name":"Performance 2",
-													"startDate":"2023-02-01",
-													"endDate":"2023-02-10"
-												  }
+											   {
+											  "id":"1",
+											  "name":"Performance 1",
+											  "startDate":"2023-01-01",
+											  "endDate":"2023-01-10"
+											   },
+											   {
+											  "id":"2",
+											  "name":"Performance 2",
+											  "startDate":"2023-02-01",
+											  "endDate":"2023-02-10"
+											   }
 											  ]
 											  """
 											  )
@@ -98,10 +48,22 @@ class LotteClientTest {
 		LotteClient lotteClient = new LotteClient(baseUrl);
 
 		// when
+		List<LotteClient.Performance> performances = lotteClient.getPerformances().getBody();
+
 		// then
-		assertThatThrownBy(() -> lotteClient.getPerformances())
-				.isInstanceOf(RuntimeException.class)
-				.hasMessage("서버 오류 발생");
+		assertThat(performances).hasSize(2)
+								.anySatisfy(performance -> {
+									assertThat(performance.getId()).isEqualTo("1");
+									assertThat(performance.getName()).isEqualTo("Performance 1");
+									assertThat(performance.getStartDate()).isEqualTo("2023-01-01");
+									assertThat(performance.getEndDate()).isEqualTo("2023-01-10");
+								})
+								.anySatisfy(performance -> {
+									assertThat(performance.getId()).isEqualTo("2");
+									assertThat(performance.getName()).isEqualTo("Performance 2");
+									assertThat(performance.getStartDate()).isEqualTo("2023-02-01");
+									assertThat(performance.getEndDate()).isEqualTo("2023-02-10");
+								});
 
 	}
